pyproject.toml
[project]
name = "mcp-capnp-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=3.4.3",
    "pycapnp>=2.0.0",
]
